[perf]: all country data moved to pinia store
