+++ src/components/Modal/CoutriesModal.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup>
 import { computed, ref, reactive } from "vue";
-import axios from "axios";
+import { onMounted } from "vue";
 import { Button } from "@/components/ui/button";
 import {
   Dialog,
@@ -12,66 +12,31 @@ import {
 } from "@/components/ui/dialog";
 import { Input } from "../ui/input";
 import { currencyToCountryMap } from "@/data";
+import { userateStore } from "@/stores/useRateStore";
+const store = userateStore();
 
-const searchQuery = ref("");
-const countriesNam = ref<{ [key: string]: any }>({});
-const selectedCountry = ref("USD");
-const isOpen = ref(false);
-const inverseRate = ref();
+onMounted(()=> {
+  store.fetchData();
 
-const fetchData = async () => {
-  try {
-    const res = await axios.get("https://www.floatrates.com/daily/ngn.json");
-    countriesNam.value = res.data;
-    console.log(countriesNam.value);
-  } catch (e) {
-    console.error("Error Fetching Data:", e);
-  }
-};
-fetchData();
-
-const filteredCountries = computed(() => {
-  if (!searchQuery.value) {
-    return Object.values(countriesNam.value);
-  }
-  return Object.values(countriesNam.value).filter((country: any) =>
-    country.name?.toLowerCase().includes(searchQuery.value.toLowerCase()),
-  );
-});
+})
 
 const getCountryCode = (currencyCode: string) => {
   return currencyToCountryMap[currencyCode.toUpperCase()] || "";
 };
 
-const objshape = {
-  aed: {
-    aplhacode: 'AED',
-    rate: 240
-  }
-}
-const getcountryRate = (countryName: string) => {
-  const lowerCaseCountryName = countryName.toLowerCase();
-  console.log(lowerCaseCountryName)
-  const countryData = countriesNam.value[lowerCaseCountryName];
-  inverseRate.value = countryData.rate
-  console.log(inverseRate.value)
-  // console.log(countriesNam.value)
-}
-
-
 const handleClick = (countrycode: string) => {
-  selectedCountry.value = countrycode;
-  getcountryRate(selectedCountry.value)
-  isOpen.value = false;
+  store.selectedCountry = countrycode;
+  store.getcountryRate(store.selectedCountry)
+  store.isOpen = false;
 };
 
 </script>
 
 <template>
-  <Dialog v-model:open="isOpen">
+  <Dialog v-model:open="store.isOpen">
     <DialogTrigger as-child>
-      <Button variant="outline" @click="isOpen = true">
-        {{ selectedCountry }}
+      <Button variant="outline" @click="store.isOpen = true">
+        {{ store.selectedCountry }}
       </Button>
     </DialogTrigger>
     <DialogContent
@@ -84,7 +49,7 @@ const handleClick = (countrycode: string) => {
         </DialogDescription>
         <div class="my-2">
           <Input
-            v-model="searchQuery"
+            v-model="store.searchQuery"
             class="max-w-4xl"
             placeholder="Search for Currency..."
             type="text"
@@ -110,7 +75,7 @@ const handleClick = (countrycode: string) => {
               {{ country.code }}
             </li>
           </ul> -->
-          <button v-for="country in filteredCountries" :key="country" class="flex items-center gap-2 mt-4" @click="handleClick(country.code)">
+          <button v-for="country in store.filteredCountries" :key="country" class="flex items-center gap-2 mt-4" @click="handleClick(country.code)">
             <span>
                 <img
                   :src="`https://flagcdn.com/w40/${getCountryCode(country.code).toLowerCase()}.png`"
